requirements.txt
seedrcc
deta
python-dotenv
feedparser
pytz
requests
pymongo
nanoid
apscheduler
humanize